moved product related functions to their own file
