+++ examples/main.go
@@ -4,5 +4,5 @@ import gosupreme "github.com/doxify/supreme-go-api"
 
 func main() {
 	s := gosupreme.New()
-	s.FetchStock()
+	s.Init()
 }
